Create Cleaning, Organizing and Presenting Data for a Business
This program will take clumsy storage of information, split, strip characters form the information, organize it into lists, present the data according to each client, get the information of every clients name, the date of the transaction, the amount of the transaction, the total daily amount. Also will print out the name, and the amount of the products that were sold that day.