Create My Encrypting Caesar Cipher
This program will Encode an specific message you have, it be a string or a list, just input the offset and the message you want to encode and as easy as that it encodes your message. Now you can chat privately with your friends and family! I guess thats nice